dom/security/test/csp/test_websocket_localhost.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Bug 1729897: Allow unsecure websocket from localhost page with CSP: upgrade-insecure </title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe style="width:100%;" id="test_ws_self_frame"></iframe>

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

function finishTest() {
  window.removeEventListener("message", receiveMessage);
  SimpleTest.finish();
}

window.addEventListener("message", receiveMessage);
function receiveMessage(event) {
  is(event.data.result, "self-ws-loaded", "websocket loaded");
  ok(event.data.url.startsWith("ws://"), `scheme must be ws:// but got ${event.data.url}`);
  finishTest();
}

SpecialPowers.pushPrefEnv({set: [
  ["network.proxy.allow_hijacking_localhost", true],
  ["network.proxy.testing_localhost_is_secure_when_hijacked", true],
]}).then(function() {
  const HOST = "http://localhost/tests/dom/security/test/csp/";
  var test_ws_self_frame = document.getElementById("test_ws_self_frame");
  test_ws_self_frame.src = HOST + "file_websocket_csp_upgrade.html";
});

</script>
</body>
</html>
